backend/requirements/base.txt
psycopg2-binary==2.9.9                  # https://github.com/psycopg/psycopg2
cryptography==43.0.0                    # https://github.com/pyca/cryptography
python-slugify==8.0.4                   # https://github.com/un33k/python-slugify
pytz==2024.2                      # https://github.com/stub42/pytz
Pillow==10.4.0                          # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0                     # https://github.com/hynek/argon2_cffi
hiredis==3.0.0                          # https://github.com/redis/hiredis-py
redis==5.1.0                            # https://github.com/redis/redis-py
celery==5.4.0                           # https://github.com/celery/celery
flower==2.0.1                           # https://github.com/mher/flower
stripe==11.1.0                          # https://github.com/stripe/stripe-python

# # Django
# # ------------------------------------------------------------------------------
django==5.1.2                           # https://www.djangoproject.com/
djangorestframework-simplejwt==5.3.1    # https://github.com/jazzband/djangorestframework-simplejwt
django-health-check==3.18.3             # https://github.com/revsys/django-health-check
django-environ==0.11.2                  # https://github.com/joke2k/django-environ
django-celery-beat==2.7.0               # https://github.com/celery/django-celery-beat
django-cors-headers==4.4.0              # https://github.com/adamchainz/django-cors-headers
djangorestframework-api-key==3.0.0      # 
jinja2==3.1.4                           # https://github.com/pallets/jinja/